sdk: rename sdkv2 to sdk
