+++ pkg/sdk/registry_test.go
