+++ pkg/testutils/node.go
@@ -19,7 +19,7 @@ import (
 	"math/rand"
 	"time"
 
-	fuddle "github.com/andydunstall/fuddle/pkg/sdkv2"
+	fuddle "github.com/andydunstall/fuddle/pkg/sdk"
 	"github.com/google/uuid"
 )
